\name{MA0004.1}
\alias{MA0003.2}
\alias{MA0004.1}
\alias{MA0048}
\alias{MA0043}
\docType{data}
\title{
  Some example PFM matrices.
}
\description{
  Some example PFM matrices from JASPAR 2014.
}
\usage{
  data(MA0004.1)
  data(MA0003.2)
  data(MA0048)
  data(MA0043)
}
\format{
  The format is: \code{PFMatrix} object.
}
\details{
  Some examples PFM matrices from JASPAR 2014.
}
\source{
  http://jaspar.genereg.net/
}
\examples{
data(MA0004.1)
data(MA0003.2)
data(MA0048)
data(MA0043)
}
\keyword{datasets}
