\name{TFBSTools-package}
\alias{TFBSTools}
\docType{package}
\title{
  TFBS software package.
}

\description{
  TFBS includes a set of tools for transcription factor binding site detection and analysis as well as database interface functions for JASPAR, etc.
}

\author{
Ge Tan
}

\keyword{package}

