\name{TFBSTools-package}
\alias{TFBSTools}
\docType{package}
\title{
  TFBS package.
}
\description{
  Tools for transcription factor binding site detection and analysis.
}
\author{
Ge Tan
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
 \code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
